fix(diffusion): lazy-preserving WanVideo Clone — don't materialize 15B params a 16GB host can't hold
WanVideoModel_Clone_CreatesIndependentCopy failed fast (35ms) on the
flat-Vector<T> int.MaxValue guard: Clone -> DiTNoisePredictor.GetParameters()
tries to allocate a single Vector of ~15.2B elements (WanVideo-14B), which both
exceeds int.MaxValue and is ~120GB at double — impossible on any 16GB CI runner.
Streaming does NOT help this path (it pages forward-time tensor data; the flat
parameter-vector materialization is still int-bounded — the guard says so).

Fix: Clone is now lazy-preserving. A never-forwarded predictor holds no resident
weights — its parameters are fully determined by its construction config — so
the freshly-constructed clonedDit (same variant config) is already an equivalent
independent copy; skip the flat GetParameters/SetParameters round-trip entirely.
When the source HAS materialized weights, copy per-layer via CopyParametersFrom
(each layer < int.MaxValue) instead of one flat aggregate.

- DiTNoisePredictor: expose AreLayersInitialized (the lazy/resolved state).
- WanVideoModel.Clone: copy weights only when AreLayersInitialized; otherwise
  produce a lazy clone in O(1).
- Test: ParameterCount is long (~15.2B); drop the (int) cast that overflowed and
  made the equality assertion meaningless at this scale.

Verified: WanVideoModel_Clone_CreatesIndependentCopy passes in 74ms (was failing).

+++ tests/AiDotNet.Tests/UnitTests/Diffusion/Models/DiffusionModelContractTests.cs
@@ -1000,7 +1000,10 @@ public class DiffusionModelContractTests : DiffusionUnitTestBase
 
         Assert.NotNull(clone);
         Assert.NotSame(model, clone);
-        Assert.Equal(model.ParameterCount, (int)clone.ParameterCount);
+        // ParameterCount is long: WanVideo-14B reports ~15.2 B params, far past
+        // int.MaxValue, so the comparison must stay in long arithmetic (a prior
+        // (int) cast overflowed and made the assertion meaningless at this scale).
+        Assert.Equal(model.ParameterCount, clone.ParameterCount);
     }
 
     [Fact(Timeout = 120000)]
